Renamed Location to LocationPage
