+++ app/containers/LocationPage/constants.js
