+++ app/containers/LocationPage/reducer.js
