+++ app/containers/LocationPage/reducer.test.js
