+++ app/containers/LocationPage/selectors.js
