+++ app/reducers.js
@@ -8,7 +8,7 @@ import { connectRouter } from 'connected-react-router';
 import history from 'utils/history';
 import globalReducer from 'containers/App/reducer';
 import languageProviderReducer from 'containers/LanguageProvider/reducer';
-import locationReducer from 'containers/Location/reducer';
+import locationReducer from 'containers/LocationPage/reducer';
 import questionnaireReducer from 'containers/QuestionnaireContainer/reducer';
 
 /**
